refactor get_read_write_buffers_sizes into BaseSchedulerNode
