+++ torch/_inductor/graph.py
@@ -915,42 +915,14 @@ class GraphLowering(torch.fx.Interpreter):
 
         scheduler = Scheduler(self.buffers)
 
-        def get_read_write_buffers_sizes(node):
-            if isinstance(node, NopKernelSchedulerNode):
-                return 0
-            reads = {dep.name for dep in node.read_writes.reads}
-            writes = {dep.name for dep in node.read_writes.writes}
-
-            def is_materialized(buf):
-                buf_uses = {user.node for user in scheduler.name_to_node[buf].users}
-                return len(buf_uses - set(node.snodes)) > 0
-
-            if isinstance(node, FusedSchedulerNode):
-                removed_buffers = {dep for dep in writes if not is_materialized(dep)}
-                writes = writes - removed_buffers
-                reads = reads - removed_buffers
-            node_bytes = 0
-            for buf in reads | writes:
-                if buf in self.name_to_buffer:
-                    buf = self.name_to_buffer[buf]
-                elif buf in self.graph_inputs:
-                    buf = self.graph_inputs[buf]
-                else:
-                    continue
-
-                node_bytes += V.graph.sizevars.size_hint(
-                    sympy_product(buf.get_size())
-                ) * get_dtype_size(buf.get_dtype())
-            return node_bytes
-
         total_bytes = 0
         node_counts = []
         node_runtimes = []
         for node in scheduler.nodes:
-            num_bytes = get_read_write_buffers_sizes(node)
+            num_bytes = node.get_read_write_buffers_sizes()
             total_bytes += num_bytes
             node_counts.append((node, num_bytes // 4))
-            node_runtimes.append((node, node.get_estimated_runtime(num_bytes)))
+            node_runtimes.append((node, node.get_estimated_runtime()))
         return total_bytes, node_counts, node_runtimes
 
     @dynamo_timed
